Fix for systems with crypt.
First off doesn't this belong in say kdecore?

Secondly the arguments should be constized (so that it's like the rest of
the crypt functions out there).

Third, I removed the rediculous || !defined(HAVE_CRYPT_H) part?  It's
uh stupid, because config.h was included nowhere, and there are systems
(namely Solaris and FreeBSD and uh apparently glibc based systems too)
that put the crypt prototype not in a non-existant crypt.h but in unistd.h

svn path=/trunk/kdebase/kdm/; revision=23740