Re: Bug #1469 From:   Mark Davies <mark@MCS.VUW.AC.NZ>
Below are patches to add support for Digital UNIX's C2 security to kcheckpass
(and kscreensaver).  Whats needed in addition to this is changes to the
configure stuff to define "HAVE_OSF_C2_PASSWD" and to link with "-lsecurity"
when appropriate.

cheers
mark

There are two more parts to this.  One is to update the screensavers, the
other kde-common/admin.  Ideally this would be imported into the 1.1
branch after being tested.

svn path=/trunk/kdebase/kcheckpass/; revision=25038