New KWin look-and-feel plugin specifically coded for laptops :)
It features a reduced height taskbar with oversized width buttons (I may make
these even wider to make them easier to hit), a minimal frame, and a
oversized shaped resize handle. Uses the system KWin style buttons, not just
because I like them but because they fit in the reduced size bar ;-)

As usual, here's a screenshot:
http://www.mosfet.org/kwinlaptop.gif

On another note, the Be style is going away and is now replaced by B2. Be
needs work so I asked Matthias Ettrich about it (he's the maintainer of that
style), and it turns out he's already using B2 :)

svn path=/trunk/kdebase/kwin/; revision=44656