If you use the k_dcop functionality of dcop, then you have to be careful to do it right.  Basically, it takes care of all the background functions for you... including process()
Now this actually compiled before since the author defined process.. but
as soon as dcopidl changed (added functions()), then this broke.

The correct way to handle this is to add a:

kshorturifilter.skel

file to the sources line.  Then, the dcopidl2cpp parser will generate the
skeleton functions for you!

svn path=/trunk/kdebase/kcontrol/; revision=45173