Highcolor version of B2. It uses a small set of automatically generated gradients and dynamic wallpaper to get a cool look on highcolor displays. It has all the features of a dedicated widget engine (adjusts to the current color scheme, fast, etc...) but with a cooler look (and more memory for the gradients - but not nearly as much as KDE pixmap themes ;-) It's designed to be as fast as possible with the least amount of pixmaps.
It's somewhat buggy. The toolbars are mostly correct but not nearly as
efficent as it could be, and it all needs to be converted to the new parent
pixmap deal instead of masks and resize handlers for transparency. On
lowcolor displays it looks the same as the normal B2, I would like to remove
lowcolor support and just launch B2 instead.

You can see a screenshot at:
http://mosfet.jorsm.com/b2highcolor-desert.gif

mosfet.org is down :P

svn path=/trunk/kdelibs/kstyles/; revision=46033