Tweaked the scrollbar drawing a bit.  I really like the two buttons on the bottom (or right) of the scrollbar.. but they look totally hosed when the scrollbar is tiny.  Also, they really aren't needed when the upper and lower buttons are only a few pixels from each other.
So I put in some heuristics to deal with this.  If the total length of
the 'bar is less then 4 times the buttons, then only two buttons are
drawn instead of the normal three.

I didn't merge this into B2 as I don't know if mosfet wants these changes
or not...

There are still some painting problems with small scrollbars where the
upper button is painted over by the slider button.

svn path=/trunk/kdelibs/kstyles/; revision=46468