The secret to job security is writing unmaintainable code.  I make a fine example with kcmnotify ;)
Some early support (or there lack of) for no-present.  The upcoming flag
in eventsrc so that users can't set it to present in a certain way.
Specifically KWin opening up a window which causes KNotify to open up a window,
which causes KWin to tell KNotify to open up a window.

I really just wanted to make a commit because I hadn't in a while.  But
it's far too late to actually write any decent code, so really all I
do is re-arrange the code I already wrote, make up a whole novel for the
commit log, then go off to sleep.  Oh, come on, I'm sure everyone else
has done it too.

svn path=/trunk/kdebase/kcontrol/; revision=50518