Changed the behavior of this internet keyword filter.
The old behavior was to only return false if the keyword was wrong...
if the filter data wasn't a URL at all, it would default to realnames
(or whatever the user-set default was).

That was fine if this was the only filter ever used and you *knew*
that the data was meant to be a query.

However, when you use this with all other filters, it breaks horribly.

Say you have a "url" of "konsole".  That's almost surely NOT a request
for a realname or a search.. it's probably an attempt to start a
program.  Unfortunately, this filter would *never* allow that.

So, the new behavior is to only accept URI that are not malformed.
Which means that if 'konsole' comes by or 'www.kde.org' comes by, then
it will pass it on and let another filter (or no filter.. it doesn't
care) take care of it.

svn path=/trunk/kdebase/kcontrol/; revision=51586