Defaulted widget color from RGB 220 220 220 to 192 192 192. Two reasons for doing that:
1) on laptop-displays it's quite common that you can't distinguish
white from very light gray (low contrasts).

2) From a point of ergonomics it's quite bad to stare at widgets
which are almost white most of the time. There is a reason for people
wanting to choose a different color than white in textprocessors for the
sheet of paper in the Background.

Because of these reasons almost all other colorscms were much less
bright than the previous one.

The fact that the colors for the widgets in Windows appear a little bit
brighter is due to the lack of gamma-correction in X. Now that
XFree4.0 supports on-the-fly-gamma-correction-control one might think about
a kcontrol-module for this kind of stuff?

svn path=/trunk/kdebase/kcontrol/; revision=57372