It was showing some ugly arrows on the tabs (Index/Search/Help).
This way it makes the tab as wider as it wants to be and we're
sure that there are no arrows. I hope there's no translation
where the tabs are very long (they were for sure not needed in
english).

To translators: Please check this for your language, and tell
me if it's needed to use a bound for the widths so that the tabs
are not extremely wider.

+  _tab->setMinimumWidth(_tab->sizeHint().width());
+  _tab->setMaximumWidth(_tab->sizeHint().width());

svn path=/trunk/kdebase/kcontrol/; revision=65675