* New XDM port from xfree86 4.0.1 	  Expecting (compilation) problems on non-linux systems.         * Made the greeter dynamically loadable         * Moved chooser and greeter to separate directories,           same for unused stuff (misc/)         * Optional auto-login, auto-re-login and password-free login         * A LOT of bug fixes and clean up (though still not enough)
The patch is NOT ready, most extended settings for the kcontrol modules
are ignored. Expect further commits in the next days.
It will also require a security audit once it is ready, as we don't want
somebody to be logged in automatically without our explicit permission,
do we? :)

GUI: the greeter (login window) look changed somewhat

svn path=/trunk/kdebase/kdm/; revision=73772