* Actually implemented many of the functions i already commited 	  the configuration module for (auto-login, auto-re-login, 	  password-less login). 	* Some bug fixes
Please look for security holes! Nobody should be allowed in without a
password, if he is not explicitly allowed to ...

HELP: dlclose()ing the greeter module causes a crash. I guess, that
some kde-libs and/or qt do not like being unloaded. Could somebody with
deeper knowledge investigate this, please?

svn path=/trunk/kdebase/kdm/; revision=74144