Hey, I'm actually making a commit ;-)
I'm making the taskbar themeable again, but noticed one thing. You can
now have a modified icon, which didn't exist when I made the original
KStyle routine. Easy fix tho: previously KStyle wanted to always draw the
icon and text for the button, now if you pass QString::null for the text
and NULL for the pixmap pointer it just draws the button. This allows the
taskbar to just use KStyle to draw the button and draw the contents itself.

svn path=/trunk/kdelibs/kstyles/; revision=80641