Added "Import..." button which launches kthemeimporter.
Note that importing gtk themes it utterly broken. Mosfet, could
you have a look at it? Is especially messes up with AA.
After applying a legacy theme and changing it back (with a restart
inbetween), I have no AA anymore although AntiAlias=true is set in
kdeglobals...

This is a showstopper for beta, I'd say.

svn path=/trunk/kdebase/kcontrol/; revision=97183