- optical fixes for read-only mode
i came to the conclusion, that QWidget should have a function setReadOnly(),
which would also influence all child widgets, like setEnabled(). the
difference between the functions is simple: a disabled widget is greyed out
and is completely dead to any input; it represents something that cannot be
changed, because it does not make sense to do so. read-only widgets on the
other hand look normal; the user just can't change their state, possibly
because he is not authorized to do so (like in the kdm setup).
some troll wants to comment on this?

svn path=/trunk/kdebase/kcontrol/; revision=108794