GUI:
Added support for configuring the panel's alignment when it's width
is set to a value of less than 100%. This makes it possible to have
kicker aligned left/center/right when horizontal,
and top/center/bottom when vertical.

The default is left aligned when KDE is set to an LTR language and
right aligned when it's set to an RTL language. When the language
setting is changed, the panel alignment will change as well until the
user makes an active choice about the panel's configuration by changing
the values in the dialog box and clicking the OK or Apply buttons.

svn path=/trunk/kdebase/kcontrol/; revision=128554