The most difficult patch to KControl ever:
+      *_rootProcess << "--nonewdcop";

But it makes danimo very happy, because now the root modules can use DCOP...

Note btw, that as Waldo pointed out on kde-devel/kde-core-devel, this option
makes it *impossible* to use KIO at the moment in the root modules, so it
might break some of those. A solution at the moment is only to use either
dcop or KIO, but not both. DCOP it is then ;-)
In the end there should of course be a solution for the KIO problem, but
Waldo is working on that I think (and hope...)

svn path=/trunk/kdebase/kcontrol/; revision=129605