X Resize and Rotate extension kcontrol module.
This allows you to, resize, rotate, reflect, and change the refresh rate
of each of the screens attached to your X display on the fly.

You need the latest CVS version of Xfree86 to use this; make sure you
build the libraries too by editing the host.def config file.

The support is still quite unrefined; the ui needs work, and kicker and
kwin should be adapted to behave nicer...

svn path=/trunk/kdebase/kcontrol/; revision=184499