CCMAIL: kwin@kde.org Fixed the misrendered shaded border. For me, the problem was fixed by redoing the layout in the shapeChange() function. Maybe the other styles have to modified in a similar way.
I'd like to reduce the width of the window border to the titlebar width only,
instead of the full window border. In the BeeToo, I obtain a similar effect by
painting the titlebar only, but the trick is revealed by the window snap zones,
and the transparent border. Is there a way to do this cleanly with current kwin?

svn path=/trunk/kdebase/kwin/; revision=255364