From BR62548 (by Maarten Thibaut): X libraries are determined by the auto* tools and placed in 2 defines: X_PRE_LIBS and X_EXTRA_LIBS.
configure.in incorrectly leaves out X_EXTRA_LIBS in a few spots, causing linking against -lX11 to fail on Solaris platforms because on this platform we need "-lsocket -lnsl" in order to compile against X11. X_EXTRA_LIBS contains these extra libraries, so we need to change the LIBS variable in configure.in in a couple of places to include $X_EXTRA_LIBS.

This patch also fixes an issue on some platforms where the ordering of the X11 libs is important. X_PRE_LIBS should always be put before '-lX11' and X_EXTRA_LIBS should come after.

svn path=/trunk/kdebase/kdesktop/lock/; revision=259211