- fix --dontlock
- use kcheckpass' new binary server mode -> be PAM-compliant
- port over kdm's conversation plugin architecture

yes, i'm late with my new features ... but as they are not really new
(only in a new place), coolo agreed (well, didn't object :) to commit
them now.

for packagers ... kdesktop_lock and kdm share the same plugins (well,
there is only one atm), so you prolly have to make a package which both
depend on.

svn path=/trunk/kdebase/kdesktop/lock/; revision=263547