there is no paticular reason to understand that commit - QSpacerItem sucks. as designer forces one to use it, designer sucks and we shouldn't use it. that simple. ;)
in this particular case it could be even a bug, not only general
suckiness: for a weird reason, when the other items (two widgets) in the
gridlayout's row are hidden, the spacer decides to expand vertically,
even though it is a horizontal spacer. additionally, it cannot be
squeezed by another stretch. setting the minimal sizes to zero prevents
it from doing both.

short version: make kdm's background settings top-aligned.

svn path=/trunk/kdebase/kcontrol/; revision=268729