I'm removing the fade stuff for the moment. It's rather difficult to do properly in my opinion (at least I haven't got it right yet).
It's hard to get the timing right so that it will fade just before the screensaver starts. Some screensavers take longer than others to start. And it's made more complicated because we use xscreensaver hacks and not just our own. If you have any ideas coolo I want to hear them :)

svn path=/trunk/kdebase/kdesktop/lock/; revision=343098