This has been on my wishlist for ... years (KDE 1?)
Style/Misc has only ever had 1 thing, the "Toolbar" so the tab should
just be Toolbar.  On top of that the Toolbar was in a group box all by
itself.  Group box's don't belong by themselves, they get lonly.  So I
removed the toolbar groupbox and the Effect groupbox, both of which
were redundant by the tab and uncessessary as they were all alone.

svn path=/trunk/kdebase/kcontrol/; revision=380687