Port to new KTimezones API. I have tried to preserve and merge the Solaris/BSD support into KTimezones...tests on those platforms are welcome.
FWIW, I am concerned that the existing logic for setting the system timezone
does not work in a sufficiently platform sensitive manner (e.g. it creates
a /etc/timezone system on Linux, which is not used by glibc AFAICS). But, I
did not feel confident to change that.

The only functional change is in the display of the current timezone: this now
has the real name of the current timezone, as well as the abbreviated name from
before.

svn path=/trunk/KDE/kdebase/kcontrol/; revision=432367