Moved libkscreensaver into workspace/kscreesaver/.
The current screensaver library is very specific for the kde desktop only.  In the future if a cross platform framework
is created then that would be well suited to go into kdelibs.

As for build issues kdeartwork (which contain many screensavers) requires kdebase to be installed already for the kwin themes so that shouldn't be a problem....

This makes a nice package for workspace/kscreensaver/.  It now has a library, two screensavers, README, and the applnk files.  New developers will be much more likely to understand how screensavers work in KDE.  It might even been a good idea to move the kcontrol module in here, but we will see.


svn path=/trunk/KDE/kdebase/workspace/; revision=478099