More use case bugs...
ALT-F2 forcibly steals focus now, but all other calls to display do not...

And following up the last commit, pressing ESC should hide and clear the field so I added a vanish() slot that does this.  It could
actually be useful to expose this slot via dbus as it's essentially a way to close the run dialog without killing the app.  The
name vanish() can be renamed if people don't like it :P


svn path=/trunk/KDE/kdebase/workspace/; revision=637035