an admittedly very lame Animator plugin. it only implement appearing (which is the only animation that Corona currently makes use of) and the effect is pretty .. mundane (it zooms the added component in to full size); and even then it's 'buggy' in that if the item has a transformation on it from inception, this will break that. but it's really just meant to be a test case to be filled in with something real over time.
it does show that it is quite simple to go from a desktop with no animations/effects to one with them. however, i do wonder about how much more information will need to passed in to easily do nice things like paint over/around the QGraphicsItem.

to see this in action, add the following to plasmarc:

[Phase]
animator=default

svn path=/trunk/KDE/kdebase/workspace/; revision=673399