welcome our now system activity dialog overlord!
thanks to the awesome hacking of tapsell on ksysguard we now have access to the widget via a library and can therefore load the process table in-process. this makes popping up that window almost -instant- and with a very, very light UI.

we still have to do something about the slightly oddly placed buttons and add the expander to show system load, etc (and maybe a button/link to open the full sysguard app?). will talk and hack with john on these parts, but colour me one very happy camper right now.

svn path=/trunk/KDE/kdebase/workspace/; revision=678267