i'm not sure what part of this was hard to understand, but evidently it wasn't communicated quite clearly enough. so let's try again:
extending the "sea of highlight()" rect to the full option.rect is not acceptable. if you want to see full width selections, do them up right. but clashing with icons, with usage bars and emphasizing negative space is not an option.

if you really, really dislike rect-behind-text-only then i encourage you to really, really come up with a solution, not something that satisfies you enough to ignore how stupidly *bad* it ends up looking.

and no, i'm not buying any of the BS 'usability' reasoning because *none* of the issues raised have *one iota* of basis in reality. this is an aesthetics issue, so do it aesthetically proper. or not at all.

CCMAIL:panel-devel@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=780192