This patch adds the ability to Kickoff to set its layout according to the direction the popup is appearing relative to the launcher button, to improve the ergonomics of kickoff when its launcher is not in the standard position.
Assumption 1: The hardcoded tab order defines a desirable priority order
Goal 1: TabBar alignment is perpendicular to direction of mouse travel
from launcher to target, to prevent mousing over non-target tabs and
potential unnecessary tab switches
Goal 2 the movie: Tabs are ordered by decreasing priority along the
mouse travel vector away from the launcher
Constraint: The search widget is different to the tabs and the footer is
of lowest priority so these should always be situated furthest away from
the origin

This gives vertical tabs when Kickoff is used in a vertical panel, and
it reorders the tabs so the Favourites tab is always closest to the
launcher, etc.

Paint code still needs fixing to paint tabs right when vertical.

Odd glitch when resizing kickoff on a vertical panel with launcher in SW
or SE position, causes the menu to slide off the screen - debug to
handle this is left in intentionally.

svn path=/trunk/KDE/kdebase/workspace/; revision=782945