Made setContainmentType public again to be able to deal with null-containments and to compile kdeutils again.
The change was introduced with r800682 and juding from the commit-log it was more something like an accident ;) If not, then it may an idea to also remove/rework the logic that was introduced with http://lists.kde.org/?l=kde-panel-devel&m=120474698213048&w=4 which allows null-containments.

p.s. if the monday-rule is not valid anymore, please check at least that things still compile or ppl may lose hours in the middle of the night to just be able to continue work on something else. Thx :)


svn path=/trunk/KDE/kdebase/workspace/; revision=800860