krdb: explicitly reference the global palette when creating the new color palette for export to non-KDE.
Before we were apparently getting an app-local palette. In the case of the color kcm changing the color scheme, this would be the previous, out of date palette (I think because the app-local config doesn't pick up the changes until it has a chance to process the signal from KGS, which hasn't happened yet when we called runRdb) instead of the one just set. This caused us to export the previous, rather than new, colors. The fix is to ask the palette for export to be created against the global config (which was just updated) rather than the app-local config (which KGS uses by default).

BUG: 164881

svn path=/trunk/KDE/kdebase/workspace/; revision=825356