org.kde.Amarok.Player is dead.  Long live MPRIS.
The modification to mpris.cpp is necessary for getting the correct track number from Amarok.  Everything else is just scrapping the now-useless Amarok Player interface support.

Note that Amarok support won't be quite right until signal sending is fixed in Amarok.

CCMAIL: panel-devel@kde.org
Is this worth backporting to the RC, or should it wait for 4.1.1 (since Amarok is only at alpha stage)?  If it needs backporting, can someone else do it?


svn path=/trunk/KDE/kdebase/workspace/; revision=832338