Make the updating of the task fast when you change desktop by caching the QIcon
reviewed by Alexis.

I had few seconds delay between i change desktop and the taskbar actually updates.
The reason was the loading of the icon which was not cached.  
The loading of the icon can be slow if the application has no icon (such as xterm) and the default X11 icon has to be loaded with KIconLoader



svn path=/trunk/KDE/kdebase/workspace/; revision=844702