Fix suspend and hibernate from kickoff launcher.
The problem here is that you need to use a KJob, as the actions
in Solid::PowerManager don't create one themselves.
This patch makes kickoff run these methods through powerdevil
if the DBus service is available when the action is triggered,
otherwise it uses the Solid API directly (no intervention from
powerdevil possible in that case). It does switch at runtime,
however -- so we should be fine in both cases.

As discussed with Kevin.

CCMAIL:plasma-devel@kde.org

svn path=/trunk/KDE/kdebase/workspace/; revision=865464