Fix the crash on exit bug in PopupApplet.
The Plasma::Dialog is initially created without a parent, but can later be
reparented to the Plasma::View. The problem with this is that when plasma
quits, the view and its children are deleted before the scene and its children.
So when the PopupApplet destructor is invoked, it tries to delete what is now
a dangling pointer.

Solve this problem by wrapping the dialog pointer in a QPointer.

svn path=/trunk/KDE/kdebase/workspace/; revision=867830