Improve extender layout for larger fontsizes:
- Make pointers to controls and battery layout in the extender members
- Set larger default sizes so at leat fontsize 12 doesn't get cut off

I'm completely in the dark why the QGraphicsLayout seems to impose the
size to the widgets, rather than the widgets to the layout. It would be
really nice if a change in the widget size (for example setting a
different text on a label) would actually make the layout adjust the
sizing). Right now, it doesn't so we have to hardcode sizes in some
places. :/

In the networkmanager applet, we're running into the same problem,
adding an item to a layout doesn't cause the widget to be resized.

Any clues on this are most welcome.

svn path=/trunk/KDE/kdebase/workspace/; revision=879285