Correctly update the extender's label when the battery is unplugged
This commit is dedicated to my beloved notebook luna, which fortunately
didn't die last night after the resident cat poured half a glass of
water into it.

I'm not a 100% sure this commit fixes the issue fully, but HAL decided
to report two batteries after some un- and replugging, so it's pretty
hard to test until that goes back to normal (maybe another glass of
water is needed!?). This patch does improve updating the extender's
label however, so there you go.

svn path=/trunk/KDE/kdebase/workspace/; revision=889762