Backport 907601 from trunk.
Don't let the parentLayoutItem member be wrong in case of a removal or if somebody destruct the layout.
In addItem the setParentLayoutItem is correctly set but not set back to 0 in removeItem or removeAt.
In addition i set it back to 0 if the layout die so parentLayoutItem of items that were inside the layout doesn't point to garbage after the deletion of the layout.

This work with 4.4 but since 4.5 is much more strict with custom layout, it crash due of an incorrect implementation.

svn path=/branches/KDE/4.2/kdebase/workspace/; revision=907603