try a bit harder to set the minimumsize
When resizing, the applet would resize to a smaller size than the
minimumSize of the GraphicsWidget inside, making the widgets overflow
outside the applet.

The minimimSize is actually set correctly this way, however, it still
overflows.

Maybe someone can have a look at what's going wrong here, I've been
debugging this with one of the Pardus developers (they have the same
problem with an applet using the tabbar and other QGraphicsWidgets).

svn path=/trunk/KDE/kdebase/workspace/; revision=912633