Fix the differences between kickoff labels and ksmserver
Update the labels but keep 'Shutdown' in kickoff menu. The
approach here seems to give a "one word" description of the
action and a little label with a phrase explaining the action
in kickoff menu, and the same phrase in ksmserver dialog.

So no need to change the word 'Shutdown' in this menu, but 
synced all other 'explaining phrases'.

BUG:179563

svn path=/trunk/KDE/kdebase/workspace/; revision=912668