Add a hack for bug 180857.
This hack makes DesktopLayout check if the applet whose geometry has changed has an applet 
handle attached. If it has, it will not do anything right away, as it would see the wrong 
geometry, but will start a zero timer to recheck the presence of the applet handle.
The timer is needed because when the applet handle is being destroyed, it first
changes the geometry and then(!) reparents the applet to the containment, so
the layout code still sees the applet handle in the appletGeometryChanged
signal.


svn path=/branches/KDE/4.2/kdebase/workspace/; revision=921244