CCMAIL: kde-core-devel@kde.org
Adding a signal to ksmserver, that notifies DBus when KDE is about to logout.
This can be useful for some applications, in that case for PowerDevil, as related to
bug #183273. PowerDevil should be notified when a logout operation is in progress, 
so that it won't trigger a suspension while the pc is actually shutting down. This behavior 
is (was) really annoying.

I'm CC-ing the list since I don't know who the maintainer of ksm is, and also,
should I backport this?

Obviously, a commit in PowerDevil for taking advantage of this is coming very soon


svn path=/trunk/KDE/kdebase/workspace/; revision=923137